py/requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
psycopg2-binary==2.8.6
asyncpg==0.21.0